src/requirements.txt
python-binance==0.7.11
tradingview-ta==3.2.3
windows-curses==2.2.0